Re-configure the regressions in the test cases
