+++ package.json
@@ -7,7 +7,8 @@
     "build": "vite build",
     "preview": "vite preview",
     "lint": "eslint . --ext .vue,.js,.jsx,.cjs,.mjs --fix --ignore-path .gitignore",
-    "format": "prettier --write src/"
+    "format": "prettier --write src/",
+    "test": "jest --verbose"
   },
   "dependencies": {
     "@intlify/unplugin-vue-i18n": "^1.5.0",
@@ -21,12 +22,15 @@
     "vue-router": "^4.2.5"
   },
   "devDependencies": {
+    "@babel/preset-env": "^7.23.6",
     "@quasar/vite-plugin": "^1.6.0",
     "@rushstack/eslint-patch": "^1.3.3",
     "@vitejs/plugin-vue": "^4.4.0",
     "@vue/eslint-config-prettier": "^8.0.0",
     "eslint": "^8.49.0",
     "eslint-plugin-vue": "^9.17.0",
+    "jest": "^29.7.0",
+    "joi": "^17.11.0",
     "prettier": "^3.0.3",
     "sass": "^1.33.0",
     "stylus": "^0.62.0",
