--- src/plugins/models/tests/data.js
@@ -1,618 +1,0 @@
-
-export const ALARMS = [
-    {
-      'asn': 1,
-      'asn_name': 'LVLT-1',
-      'asn_country': 'United States',
-      'asn_country_iso_code2': 'US',
-      'asn_country_iso_code3': 'USA',
-      'moas_asn_attacker': [
-        1
-      ],
-      'moas_asn_attacker_name': [
-        'LVLT-1'
-      ],
-      'moas_asn_attacker_country': [
-        'United States'
-      ],
-      'moas_asn_attacker_country_iso_code2': [
-        'US'
-      ],
-      'moas_asn_attacker_country_iso_code3': [
-        'USA'
-      ],
-      'moas_asn_attacker_af': [
-        '4, 4'
-      ],
-      'moas_asn_attacker_newcomer': [
-        true
-      ],
-      'moas_asn_attacker_attacking_prefix': [
-        '103.28.85.0/24, 103.98.129.0/24'
-      ],
-      'moas_asn_attacker_trustworthy': [
-        true
-      ],
-      'moas_asn_attacker_trustworthy_tag': [
-        'all-newcomers-next-to-an-oldcomer, newcomer-small-asn, rpki-all-newcomer-invalid-roa, rpki-some-newcomer-invalid-roa, not-previously-announced-by-any-newcomer'
-      ],
-      'moas_asn_victim': [
-        58504
-      ],
-      'moas_asn_victim_name': [
-        'TECHMINDS-NP TECHMINDS NETWORKS PVT. LTD.'
-      ],
-      'moas_asn_victim_country': [
-        'Nepal'
-      ],
-      'moas_asn_victim_country_iso_code2': [
-        'NP'
-      ],
-      'moas_asn_victim_country_iso_code3': [
-        'NPL'
-      ],
-      'moas_asn_victim_newcomer': [
-        false
-      ],
-      'moas_asn_victim_trustworthy': [
-        true
-      ],
-      'moas_asn_victim_trustworthy_tag': [
-        'all-newcomers-next-to-an-oldcomer, newcomer-small-asn, rpki-all-newcomer-invalid-roa, rpki-some-newcomer-invalid-roa, not-previously-announced-by-any-newcomer'
-      ],
-      'moas_asn_victim_af': [
-        '4, 4'
-      ],
-      'moas_timebin': [
-        1697784625
-      ],
-      'moas_count': [
-        1
-      ],
-      'moas_severity': [
-        'high'
-      ],
-      'moas_deviation': [
-        80
-      ],
-      'moas_duration': [
-        5
-      ],
-      'moas_key': 'asn_attacker',
-      'asn_name_truncated': 'LVLT-1 (AS1)',
-      'bgp_entity': [
-        1,
-        1
-      ],
-      'bgp_entity_type': [
-        'asn',
-        'asn'
-      ],
-      'bgp_entity_name': [
-        'LVLT-1',
-        'LVLT-1'
-      ],
-      'bgp_entity_country': [
-        'United States',
-        'United States'
-      ],
-      'bgp_entity_country_iso_code2': [
-        'US',
-        'US'
-      ],
-      'bgp_entity_country_iso_code3': [
-        'USA',
-        'USA'
-      ],
-      'bgp_entity_af': [
-        '4, 6',
-        '4, 6'
-      ],
-      'bgp_entity_ip_count': [
-        11416,
-        11416
-      ],
-      'bgp_entity_alarm_type': [
-        'bgp',
-        'bgp'
-      ],
-      'bgp_condition': [
-        'normal',
-        '< 0.99'
-      ],
-      'bgp_value': [
-        58,
-        56
-      ],
-      'bgp_historical_value': [
-        57,
-        57
-      ],
-      'bgp_severity': [
-        'medium',
-        'high'
-      ],
-      'bgp_timebin': [
-        1697781600,
-        1697781900
-      ],
-      'bgp_count': [
-        1,
-        1
-      ],
-      'bgp_key': 'entity'
-    },
-    {
-      'asn': 2,
-      'asn_name': 'UDEL-DCN',
-      'asn_country': 'United States',
-      'asn_country_iso_code2': 'US',
-      'asn_country_iso_code3': 'USA',
-      'moas_asn_attacker': [
-        2
-      ],
-      'moas_asn_attacker_name': [
-        'UDEL-DCN'
-      ],
-      'moas_asn_attacker_country': [
-        'United States'
-      ],
-      'moas_asn_attacker_country_iso_code2': [
-        'US'
-      ],
-      'moas_asn_attacker_country_iso_code3': [
-        'USA'
-      ],
-      'moas_asn_attacker_af': [
-        4
-      ],
-      'moas_asn_attacker_newcomer': [
-        true
-      ],
-      'moas_asn_attacker_attacking_prefix': [
-        '103.120.237.0/24'
-      ],
-      'moas_asn_attacker_trustworthy': [
-        true
-      ],
-      'moas_asn_attacker_trustworthy_tag': [
-        'rpki-all-newcomer-invalid-roa, rpki-all-oldcomer-invalid-roa, rpki-some-newcomer-invalid-roa, origin-small-edit-distance, rpki-some-oldcomer-invalid-roa, not-previously-announced-by-any-newcomer'
-      ],
-      'moas_asn_victim': [
-        4
-      ],
-      'moas_asn_victim_name': [
-        'ISI-AS'
-      ],
-      'moas_asn_victim_country': [
-        'United States'
-      ],
-      'moas_asn_victim_country_iso_code2': [
-        'US'
-      ],
-      'moas_asn_victim_country_iso_code3': [
-        'USA'
-      ],
-      'moas_asn_victim_newcomer': [
-        false
-      ],
-      'moas_asn_victim_trustworthy': [
-        true
-      ],
-      'moas_asn_victim_trustworthy_tag': [
-        'rpki-all-newcomer-invalid-roa, rpki-all-oldcomer-invalid-roa, rpki-some-newcomer-invalid-roa, origin-small-edit-distance, rpki-some-oldcomer-invalid-roa, not-previously-announced-by-any-newcomer'
-      ],
-      'moas_asn_victim_af': [
-        4
-      ],
-      'moas_timebin': [
-        1697794226
-      ],
-      'moas_count': [
-        1
-      ],
-      'moas_severity': [
-        'high'
-      ],
-      'moas_deviation': [
-        80
-      ],
-      'moas_duration': [
-        45
-      ],
-      'moas_key': 'asn_attacker',
-      'asn_name_truncated': 'UDEL-DCN (AS2)'
-    },
-    {
-      'asn': 27717,
-      'asn_name': 'Corporacion Digitel C.A.',
-      'asn_country': 'Venezuela',
-      'asn_country_iso_code2': 'VE',
-      'asn_country_iso_code3': 'VEN',
-      'network_delay_startpoint': [
-        27717,
-        27717
-      ],
-      'network_delay_startpoint_type': [
-        'AS',
-        'AS'
-      ],
-      'network_delay_startpoint_name': [
-        'Corporacion Digitel C.A.',
-        'Corporacion Digitel C.A.'
-      ],
-      'network_delay_startpoint_country': [
-        'Venezuela',
-        'Venezuela'
-      ],
-      'network_delay_startpoint_country_iso_code2': [
-        'VE',
-        'VE'
-      ],
-      'network_delay_startpoint_country_iso_code3': [
-        'VEN',
-        'VEN'
-      ],
-      'network_delay_startpoint_af': [
-        4,
-        4
-      ],
-      'network_delay_endpoint': [
-        3320,
-        3320
-      ],
-      'network_delay_endpoint_type': [
-        'AS',
-        'AS'
-      ],
-      'network_delay_endpoint_name': [
-        'DTAG Deutsche Telekom AG',
-        'DTAG Deutsche Telekom AG'
-      ],
-      'network_delay_endpoint_country': [
-        'Germany',
-        'Germany'
-      ],
-      'network_delay_endpoint_country_iso_code2': [
-        'DE',
-        'DE'
-      ],
-      'network_delay_endpoint_country_iso_code3': [
-        'DEU',
-        'DEU'
-      ],
-      'network_delay_endpoint_af': [
-        4,
-        4
-      ],
-      'network_delay_count': [
-        1,
-        1
-      ],
-      'network_delay_timebin': [
-        1697829300,
-        1697829300
-      ],
-      'network_delay_severity': [
-        'low',
-        'low'
-      ],
-      'network_delay_deviation': [
-        10.0817387384346,
-        10.6897085814724
-      ],
-      'network_delay_key': 'startpoint',
-      'asn_name_truncated': 'Corporacio (AS27717)'
-    }
-  ]
-  
-  export const ALARM_TYPES_MAP = {
-    'hegemony': 'AS Dependency',
-    'network_delay': 'Network Delay',
-    'network_disconnection': 'Network Disconnection',
-    'moas': 'MOAS',
-    'submoas': 'Sub-MOAS',
-    'defcon': 'DEFCON',
-    'edges': 'Fake Path',
-    'ping_slash24': 'Ping',
-    'bgp': 'BGP',
-    'ucsd_nt': 'UCSD Telescope',
-    'merit_nt': 'Merit Telescope'
-  }
-  
-  export const IHR_ALARMS_MOCKED = [
-    {
-      'timebin': '2023-10-20T23:45:00Z',
-      'originasn': 197150,
-      'asn': 12418,
-      'deviation': 35.2348037171613,
-      'af': 4,
-      'asn_name': 'QUANTUM Quantum CJSC, RU',
-      'originasn_name': 'MCRBN-AS Federal State Unitary Enterprise of the Order of the Red Banner of Labour "Russian Broad-casting and Notification Network", RU',
-      'event_type': 'hegemony'
-    },
-    {
-      'timebin': '2023-10-20T23:15:00Z',
-      'originasn': 135647,
-      'asn': 45355,
-      'deviation': 22.9060399467486,
-      'af': 4,
-      'asn_name': 'DIGICELPACIFIC-1-AP Digicel Fiji Limited, FJ',
-      'originasn_name': 'AFL-AS-AP Airports Fiji Limited, FJ',
-      'event_type': 'hegemony'
-    },
-    {
-      'timebin': '2023-10-20T23:15:00Z',
-      'startpoint_type': 'AS',
-      'startpoint_name': '8866',
-      'startpoint_af': 4,
-      'endpoint_type': 'AS',
-      'endpoint_name': '42473',
-      'endpoint_af': 4,
-      'deviation': 11.9389361531127,
-      'event_type': 'network_delay'
-    },
-    {
-      'timebin': '2023-10-20T23:15:00Z',
-      'startpoint_type': 'AS',
-      'startpoint_name': '397143',
-      'startpoint_af': 4,
-      'endpoint_type': 'AS',
-      'endpoint_name': '6461',
-      'endpoint_af': 4,
-      'deviation': 14.6861362079549,
-      'event_type': 'network_delay'
-    },
-    {
-      'id': 131964651,
-      'streamtype': 'asn',
-      'streamname': '58224',
-      'starttime': '2023-10-20T19:07:00Z',
-      'endtime': '2023-10-20T19:07:00Z',
-      'avglevel': 12,
-      'nbdiscoprobes': 10,
-      'totalprobes': 32,
-      'ongoing': true,
-      'discoprobes': [
-        {
-          'probe_id': 27080,
-          'ipv4': '151.245.9.12',
-          'prefixv4': '151.245.8.0/22',
-          'event': 131964651,
-          'starttime': '2023-10-20T19:07:24Z',
-          'endtime': '2023-10-20T19:07:24Z',
-          'level': 12,
-          'lat': 37.2875,
-          'lon': 49.6015
-        },
-        {
-          'probe_id': 26337,
-          'ipv4': '151.245.9.12',
-          'prefixv4': '151.245.8.0/22',
-          'event': 131964651,
-          'starttime': '2023-10-20T19:07:02Z',
-          'endtime': '2023-10-20T19:07:02Z',
-          'level': 12,
-          'lat': 37.2705,
-          'lon': 49.5785
-        },
-        {
-          'probe_id': 26313,
-          'ipv4': '151.245.9.12',
-          'prefixv4': '151.245.8.0/22',
-          'event': 131964651,
-          'starttime': '2023-10-20T19:07:02Z',
-          'endtime': '2023-10-20T19:07:02Z',
-          'level': 12,
-          'lat': 37.2685,
-          'lon': 49.5575
-        },
-        {
-          'probe_id': 24749,
-          'ipv4': '151.245.9.12',
-          'prefixv4': '151.245.8.0/22',
-          'event': 131964651,
-          'starttime': '2023-10-20T19:07:01Z',
-          'endtime': '2023-10-20T19:07:01Z',
-          'level': 12,
-          'lat': 37.2915,
-          'lon': 49.5695
-        },
-        {
-          'probe_id': 26351,
-          'ipv4': '151.245.9.12',
-          'prefixv4': '151.245.8.0/22',
-          'event': 131964651,
-          'starttime': '2023-10-20T19:02:34Z',
-          'endtime': '2023-10-20T19:02:34Z',
-          'level': 12,
-          'lat': 37.2775,
-          'lon': 49.5995
-        },
-        {
-          'probe_id': 26328,
-          'ipv4': '151.245.9.12',
-          'prefixv4': '151.245.8.0/22',
-          'event': 131964651,
-          'starttime': '2023-10-20T19:07:01Z',
-          'endtime': '2023-10-20T19:07:01Z',
-          'level': 12,
-          'lat': 37.2715,
-          'lon': 49.5885
-        },
-        {
-          'probe_id': 24801,
-          'ipv4': '151.245.9.12',
-          'prefixv4': '151.245.8.0/22',
-          'event': 131964651,
-          'starttime': '2023-10-20T19:07:01Z',
-          'endtime': '2023-10-20T19:07:01Z',
-          'level': 12,
-          'lat': 37.2675,
-          'lon': 49.5575
-        },
-        {
-          'probe_id': 24890,
-          'ipv4': '151.245.9.12',
-          'prefixv4': '151.245.8.0/22',
-          'event': 131964651,
-          'starttime': '2023-10-20T19:07:01Z',
-          'endtime': '2023-10-20T19:07:01Z',
-          'level': 12,
-          'lat': 37.2675,
-          'lon': 49.5915
-        },
-        {
-          'probe_id': 24806,
-          'ipv4': '151.245.9.12',
-          'prefixv4': '151.245.8.0/22',
-          'event': 131964651,
-          'starttime': '2023-10-20T19:07:00Z',
-          'endtime': '2023-10-20T19:07:00Z',
-          'level': 12,
-          'lat': 37.2705,
-          'lon': 49.5875
-        },
-        {
-          'probe_id': 25072,
-          'ipv4': '151.245.9.12',
-          'prefixv4': '151.245.8.0/22',
-          'event': 131964651,
-          'starttime': '2023-10-20T19:07:01Z',
-          'endtime': '2023-10-20T19:07:01Z',
-          'level': 12,
-          'lat': 29.5485,
-          'lon': 52.5905
-        }
-      ],
-      'duration': null,
-      'event_type': 'network_disconnection'
-    },
-    {
-      'id': 131964645,
-      'streamtype': 'country',
-      'streamname': 'IR',
-      'starttime': '2023-10-20T19:07:00Z',
-      'endtime': '2023-10-20T19:21:33Z',
-      'avglevel': 11,
-      'nbdiscoprobes': 10,
-      'totalprobes': 102,
-      'ongoing': false,
-      'discoprobes': [
-        {
-          'probe_id': 27080,
-          'ipv4': '151.245.9.12',
-          'prefixv4': '151.245.8.0/22',
-          'event': 131964645,
-          'starttime': '2023-10-20T19:07:24Z',
-          'endtime': '2023-10-20T19:12:57Z',
-          'level': 11,
-          'lat': 37.2875,
-          'lon': 49.6015
-        },
-        {
-          'probe_id': 26351,
-          'ipv4': '151.245.9.12',
-          'prefixv4': '151.245.8.0/22',
-          'event': 131964645,
-          'starttime': '2023-10-20T19:02:34Z',
-          'endtime': '2023-10-20T19:16:16Z',
-          'level': 11,
-          'lat': 37.2775,
-          'lon': 49.5995
-        },
-        {
-          'probe_id': 26337,
-          'ipv4': '151.245.9.12',
-          'prefixv4': '151.245.8.0/22',
-          'event': 131964645,
-          'starttime': '2023-10-20T19:07:02Z',
-          'endtime': '2023-10-20T19:07:02Z',
-          'level': 11,
-          'lat': 37.2705,
-          'lon': 49.5785
-        },
-        {
-          'probe_id': 26313,
-          'ipv4': '151.245.9.12',
-          'prefixv4': '151.245.8.0/22',
-          'event': 131964645,
-          'starttime': '2023-10-20T19:07:02Z',
-          'endtime': '2023-10-20T19:07:02Z',
-          'level': 11,
-          'lat': 37.2685,
-          'lon': 49.5575
-        },
-        {
-          'probe_id': 25072,
-          'ipv4': '151.245.9.12',
-          'prefixv4': '151.245.8.0/22',
-          'event': 131964645,
-          'starttime': '2023-10-20T19:07:01Z',
-          'endtime': '2023-10-20T19:23:18Z',
-          'level': 11,
-          'lat': 29.5485,
-          'lon': 52.5905
-        },
-        {
-          'probe_id': 24890,
-          'ipv4': '151.245.9.12',
-          'prefixv4': '151.245.8.0/22',
-          'event': 131964645,
-          'starttime': '2023-10-20T19:07:01Z',
-          'endtime': '2023-10-20T19:25:40Z',
-          'level': 11,
-          'lat': 37.2675,
-          'lon': 49.5915
-        },
-        {
-          'probe_id': 24801,
-          'ipv4': '151.245.9.12',
-          'prefixv4': '151.245.8.0/22',
-          'event': 131964645,
-          'starttime': '2023-10-20T19:07:01Z',
-          'endtime': '2023-10-20T19:19:48Z',
-          'level': 11,
-          'lat': 37.2675,
-          'lon': 49.5575
-        },
-        {
-          'probe_id': 26328,
-          'ipv4': '151.245.9.12',
-          'prefixv4': '151.245.8.0/22',
-          'event': 131964645,
-          'starttime': '2023-10-20T19:07:01Z',
-          'endtime': '2023-10-20T19:07:01Z',
-          'level': 11,
-          'lat': 37.2715,
-          'lon': 49.5885
-        },
-        {
-          'probe_id': 24749,
-          'ipv4': '151.245.9.12',
-          'prefixv4': '151.245.8.0/22',
-          'event': 131964645,
-          'starttime': '2023-10-20T19:07:01Z',
-          'endtime': '2023-10-20T19:27:41Z',
-          'level': 11,
-          'lat': 37.2915,
-          'lon': 49.5695
-        },
-        {
-          'probe_id': 24806,
-          'ipv4': '151.245.9.12',
-          'prefixv4': '151.245.8.0/22',
-          'event': 131964645,
-          'starttime': '2023-10-20T19:07:00Z',
-          'endtime': '2023-10-20T19:07:00Z',
-          'level': 11,
-          'lat': 37.2705,
-          'lon': 49.5875
-        }
-      ],
-      'duration': 15,
-      'event_type': 'network_disconnection'
-    }
-  ]
-  
-
